Вынос функции скролла до элемента в отдельную директиву
