+++ src/utils/ScrollToElement/ScrollToElement.ts
@@ -1,0 +1,8 @@
+export const scrollToElement = (e: React.MouseEvent<HTMLElement>, link: string) => {
+    e.preventDefault();
+    const el = document.getElementById(link);
+    el?.scrollIntoView({
+        behavior: 'smooth',
+        block: 'start',
+    });
+};
